+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - sorting the list_task tool result tasks by descending updated date
 - added missing gap between messages in virtualized message rendering
 - added support for infinite agent iterations
+- correctly installing repo-root extensions
 
 ## [0.65.0]